+++ src/index.ts
@@ -13,6 +13,8 @@ export * from './types/types';
 
 export * from './types/nft-types';
 
+export * from './types/prices-types';
+
 export { Alchemy } from './api/alchemy';
 
 export { Wallet } from './api/alchemy-wallet';
@@ -38,6 +40,8 @@ export type { TransactNamespace } from './api/transact-namespace';
 export type { NotifyNamespace } from './api/notify-namespace';
 
 export type { DebugNamespace } from './api/debug-namespace';
+
+export type { PricesNamespace } from './api/prices-namespace';
 
 export { fromHex, toHex, isHex } from './api/util';